# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 청년일자리도약장려금 관리 시스템 API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  manager: "manager",
} as const;

export interface User {
  id: number;
  username: string;
  name: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthToken {
  token: string;
  user: User;
}

export type UserInputRole = (typeof UserInputRole)[keyof typeof UserInputRole];

export const UserInputRole = {
  admin: "admin",
  manager: "manager",
} as const;

export interface UserInput {
  username: string;
  password: string;
  name: string;
  role: UserInputRole;
}

export interface UserUpdate {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  password?: string | null;
  /** @nullable */
  role?: string | null;
}

export interface Business {
  id: number;
  name: string;
  representativeName: string;
  businessNumber: string;
  /** @nullable */
  foundedDate?: string | null;
  /** @nullable */
  representativeResidentNumber?: string | null;
  /** @nullable */
  representativePhone?: string | null;
  /** @nullable */
  certPassword?: string | null;
  managerName: string;
  managerId: number;
  /** @nullable */
  operatingAgency?: string | null;
  /** @nullable */
  operatingAgencyPhone?: string | null;
  applicationYear: number;
  /** @nullable */
  bankName?: string | null;
  /** @nullable */
  accountNumber?: string | null;
  status: string;
  /** @nullable */
  workerCount?: number | null;
  createdAt: string;
}

export interface BusinessInput {
  name: string;
  representativeName: string;
  businessNumber: string;
  /** @nullable */
  foundedDate?: string | null;
  /** @nullable */
  representativeResidentNumber?: string | null;
  /** @nullable */
  representativePhone?: string | null;
  /** @nullable */
  certPassword?: string | null;
  managerName: string;
  managerId: number;
  /** @nullable */
  operatingAgency?: string | null;
  /** @nullable */
  operatingAgencyPhone?: string | null;
  applicationYear: number;
  /** @nullable */
  bankName?: string | null;
  /** @nullable */
  accountNumber?: string | null;
  status: string;
}

export interface BusinessUpdate {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  representativeName?: string | null;
  /** @nullable */
  businessNumber?: string | null;
  /** @nullable */
  foundedDate?: string | null;
  /** @nullable */
  representativeResidentNumber?: string | null;
  /** @nullable */
  representativePhone?: string | null;
  /** @nullable */
  certPassword?: string | null;
  /** @nullable */
  managerName?: string | null;
  /** @nullable */
  managerId?: number | null;
  /** @nullable */
  operatingAgency?: string | null;
  /** @nullable */
  operatingAgencyPhone?: string | null;
  /** @nullable */
  applicationYear?: number | null;
  /** @nullable */
  bankName?: string | null;
  /** @nullable */
  accountNumber?: string | null;
  /** @nullable */
  status?: string | null;
}

export type WorkerStatus = (typeof WorkerStatus)[keyof typeof WorkerStatus];

export const WorkerStatus = {
  active: "active",
  resigned: "resigned",
  completed: "completed",
} as const;

export interface Worker {
  id: number;
  businessId: number;
  /** @nullable */
  businessName?: string | null;
  name: string;
  hireDate: string;
  /** @nullable */
  residentNumber?: string | null;
  isMidJoiner: boolean;
  status: WorkerStatus;
  /** @nullable */
  resignDate?: string | null;
  /** @nullable */
  resignReason?: string | null;
  /** @nullable */
  canApplyAfterResign?: boolean | null;
  createdAt: string;
}

export interface WorkerInput {
  businessId: number;
  name: string;
  hireDate: string;
  /** @nullable */
  residentNumber?: string | null;
  isMidJoiner: boolean;
  status?: string;
}

export interface WorkerUpdate {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  hireDate?: string | null;
  /** @nullable */
  residentNumber?: string | null;
  /** @nullable */
  isMidJoiner?: boolean | null;
  /** @nullable */
  status?: string | null;
}

export type SubsidyRoundStatus =
  (typeof SubsidyRoundStatus)[keyof typeof SubsidyRoundStatus];

export const SubsidyRoundStatus = {
  scheduled: "scheduled",
  applied: "applied",
  paid: "paid",
  not_applied: "not_applied",
  resigned: "resigned",
  not_applicable: "not_applicable",
} as const;

export interface SubsidyRound {
  id: number;
  workerId: number;
  businessId: number;
  /** @nullable */
  workerName?: string | null;
  /** @nullable */
  businessName?: string | null;
  /** @nullable */
  managerName?: string | null;
  roundNumber: number;
  dueDate: string;
  amount: number;
  status: SubsidyRoundStatus;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  round2NotApplied?: boolean | null;
  createdAt: string;
}

export interface WorkerDetail {
  id: number;
  businessId: number;
  /** @nullable */
  businessName?: string | null;
  name: string;
  hireDate: string;
  /** @nullable */
  residentNumber?: string | null;
  isMidJoiner: boolean;
  status: string;
  /** @nullable */
  resignDate?: string | null;
  /** @nullable */
  resignReason?: string | null;
  /** @nullable */
  canApplyAfterResign?: boolean | null;
  createdAt: string;
  rounds: SubsidyRound[];
}

export interface ResignInput {
  resignDate: string;
  resignReason: string;
  canApplyAfterResign?: boolean;
}

export interface RoundInput {
  workerId: number;
  roundNumber: number;
  /** @nullable */
  amount?: number | null;
}

export interface RoundUpdate {
  /** @nullable */
  status?: string | null;
  /** @nullable */
  amount?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface Notification {
  id: number;
  roundId: number;
  workerId: number;
  businessId: number;
  /** @nullable */
  workerName?: string | null;
  /** @nullable */
  businessName?: string | null;
  type: string;
  /** @nullable */
  message?: string | null;
  sentAt: string;
  isRead: boolean;
}

export interface SlackSendResult {
  sent: number;
  total: number;
  message: string;
}

export interface YearCount {
  year: number;
  count: number;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface DashboardSummary {
  totalBusinesses: number;
  totalWorkers: number;
  totalRounds: number;
  pendingRounds: number;
  completedThisMonth: number;
  feeThisMonth: number;
  overdueRounds: number;
  upcomingThisWeek: number;
  businessesByYear: YearCount[];
  roundsByStatus: StatusCount[];
}

export interface ManagerStat {
  managerId: number;
  managerName: string;
  businessCount: number;
  workerCount: number;
  pendingRounds: number;
  completedRounds: number;
}

export type ListBusinessesParams = {
  /**
   * @nullable
   */
  year?: number | null;
  /**
   * @nullable
   */
  managerId?: number | null;
  /**
   * @nullable
   */
  search?: string | null;
  /**
   * @nullable
   */
  status?: string | null;
};

export type ListWorkersParams = {
  businessId: number;
  /**
   * @nullable
   */
  status?: string | null;
  /**
   * @nullable
   */
  search?: string | null;
  /**
   * @nullable
   */
  round?: number | null;
};

export type ListRoundsParams = {
  /**
   * @nullable
   */
  workerId?: number | null;
  /**
   * @nullable
   */
  businessId?: number | null;
  /**
   * @nullable
   */
  status?: string | null;
  /**
   * @nullable
   */
  roundNumber?: number | null;
  /**
   * @nullable
   */
  dueSoon?: boolean | null;
};

export type ListNotificationsParams = {
  /**
   * @nullable
   */
  unreadOnly?: boolean | null;
};