# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: 청년일자리도약장려금 관리 시스템 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthToken"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /businesses:
    get:
      operationId: listBusinesses
      tags: [businesses]
      summary: List businesses
      parameters:
        - name: year
          in: query
          schema:
            type: ["integer", "null"]
        - name: managerId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of businesses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Business"
    post:
      operationId: createBusiness
      tags: [businesses]
      summary: Create a new business
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessInput"
      responses:
        "201":
          description: Created business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
  /businesses/{id}:
    get:
      operationId: getBusiness
      tags: [businesses]
      summary: Get a business by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
    patch:
      operationId: updateBusiness
      tags: [businesses]
      summary: Update a business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessUpdate"
      responses:
        "200":
          description: Updated business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
    delete:
      operationId: deleteBusiness
      tags: [businesses]
      summary: Delete a business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Workers — use flat query params only (no nested query params to avoid Orval collision)
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthToken:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, username, name, role, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, manager]
        createdAt:
          type: string
...
    Business:
      type: object
      required: [id, name, representativeName, businessNumber, applicationYear, managerId, managerName, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        representativeName:
          type: string
        businessNumber:
          type: string
        foundedDate:
          type: ["string", "null"]
        representativeResidentNumber:
          type: ["string", "null"]
        representativePhone:
          type: ["string", "null"]
        certPassword:
          type: ["string", "null"]
        managerName:
          type: string
        managerId:
          type: integer
        operatingAgency:
          type: ["string", "null"]
        operatingAgencyPhone:
          type: ["string", "null"]
        applicationYear:
          type: integer
        bankName:
          type: ["string", "null"]
        accountNumber:
          type: ["string", "null"]
        status:
          type: string
        workerCount:
          type: ["integer", "null"]
        createdAt:
          type: string
    BusinessInput:
      type: object
      required: [name, representativeName, businessNumber, applicationYear, managerId, managerName, status]
      properties:
        name:
          type: string
        representativeName:
          type: string
        businessNumber:
          type: string
        foundedDate:
          type: ["string", "null"]
        representativeResidentNumber:
          type: ["string", "null"]
        representativePhone:
          type: ["string", "null"]
        certPassword:
          type: ["string", "null"]
        managerName:
          type: string
        managerId:
          type: integer
        operatingAgency:
          type: ["string", "null"]
        operatingAgencyPhone:
          type: ["string", "null"]
        applicationYear:
          type: integer
        bankName:
          type: ["string", "null"]
        accountNumber:
          type: ["string", "null"]
        status:
          type: string
    BusinessUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        representativeName:
          type: ["string", "null"]
        businessNumber:
          type: ["string", "null"]
        foundedDate:
          type: ["string", "null"]
        representativeResidentNumber:
          type: ["string", "null"]
        representativePhone:
          type: ["string", "null"]
        certPassword:
          type: ["string", "null"]
        managerName:
          type: ["string", "null"]
        managerId:
          type: ["integer", "null"]
        operatingAgency:
          type: ["string", "null"]
        operatingAgencyPhone:
          type: ["string", "null"]
        applicationYear:
          type: ["integer", "null"]
        bankName:
          type: ["string", "null"]
        accountNumber:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
...
    Notification:
      type: object
      required: [id, roundId, workerId, businessId, type, sentAt, isRead]
      properties:
        id:
          type: integer
        roundId:
          type: integer
        workerId:
          type: integer
        businessId:
          type: integer
        workerName:
          type: ["string", "null"]
        businessName:
          type: ["string", "null"]
        type:
          type: string
        message:
          type: ["string", "null"]
        sentAt:
          type: string
        isRead:
          type: boolean
...