# lib/api-zod/src/generated/types/businessInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 청년일자리도약장려금 관리 시스템 API
 * OpenAPI spec version: 0.1.0
 */

export interface BusinessInput {
  name: string;
  representativeName: string;
  businessNumber: string;
  /** @nullable */
  foundedDate?: string | null;
  /** @nullable */
  representativeResidentNumber?: string | null;
  /** @nullable */
  representativePhone?: string | null;
  /** @nullable */
  certPassword?: string | null;
  managerName: string;
  managerId: number;
  /** @nullable */
  operatingAgency?: string | null;
  /** @nullable */
  operatingAgencyPhone?: string | null;
  applicationYear: number;
  /** @nullable */
  bankName?: string | null;
  /** @nullable */
  accountNumber?: string | null;
  status: string;
}